refactoring of the casting, went for doubles on herons formula calculation so 1 cast needs to happen instead of 4

+++ Program.cs
@@ -211,7 +211,7 @@
 
                 else
                 {
-                    decimal sideA;
+                    double sideA;
                     decimal sideB;
                     Console.WriteLine("Enter your 3 sides of your triangle");
                     do
